Reworked the consume mechanism to work better
This change was needed for slowly growing files coming in from
scanner. The old implementation grabbed them even when they were
not completely written.
So we wait now (when not using inotify) and watch for changes in
mtime and/or size of the file. Normally this gives enough time to
the scanning process to complete the work and transfer.