Refactored the directory/file handling
There were different functions to cleanup the directory structure
in MEDIAROOT/originals when using a intelligent file naming format.
This was reduced and cleaned up to be more to the point. The idea is
to address all situations, when a file is moved or could be renamed
by us or others. After this it is time to clean up empty dirs.
There is no need for optimization, so one recursive function can do
this for us.